# schemas/v1/interface-api.openapi.json
{
  "components": {
    "schemas": {
...
      "CapabilitiesView": {
        "additionalProperties": false,
        "properties": {
          "hardware_actions": {
            "const": false,
            "title": "Hardware Actions",
            "type": "boolean"
          },
          "read_only": {
            "const": true,
            "title": "Read Only",
            "type": "boolean"
          }
        },
        "required": [
          "read_only",
          "hardware_actions"
        ],
        "title": "CapabilitiesView",
        "type": "object"
      },
      "ComparisonAreaSummaryView": {
        "additionalProperties": false,
        "properties": {
          "added": {
            "title": "Added",
            "type": "integer"
          },
          "area": {
            "enum": [
              "stage",
              "requirement",
              "feature",
              "report",
              "check",
              "finding",
              "profile",
              "artifact",
              "package"
            ],
            "title": "Area",
            "type": "string"
          },
          "changed": {
            "title": "Changed",
            "type": "integer"
          },
          "removed": {
            "title": "Removed",
            "type": "integer"
          }
        },
        "required": [
          "area",
          "added",
          "removed",
          "changed"
        ],
        "title": "ComparisonAreaSummaryView",
        "type": "object"
      },
      "ComparisonChangeView": {
        "additionalProperties": false,
        "properties": {
          "area": {
            "enum": [
              "stage",
              "requirement",
              "feature",
              "report",
              "check",
              "finding",
              "profile",
              "artifact",
              "package"
            ],
            "title": "Area",
            "type": "string"
          },
          "base_value": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Value"
          },
          "boundary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Boundary"
          },
          "candidate_value": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Candidate Value"
          },
          "change": {
            "enum": [
              "added",
              "removed",
              "changed"
            ],
            "title": "Change",
            "type": "string"
          },
          "detail_complete": {
            "title": "Detail Complete",
            "type": "boolean"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "record_key": {
            "title": "Record Key",
            "type": "string"
          }
        },
        "required": [
          "area",
          "record_key",
          "label",
          "change",
          "base_value",
          "candidate_value",
          "detail_complete",
          "boundary"
        ],
        "title": "ComparisonChangeView",
        "type": "object"
      },
      "ComparisonRevisionView": {
        "additionalProperties": false,
        "properties": {
          "achieved_evidence_level": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Achieved Evidence Level"
          },
          "allowed_claim": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Allowed Claim"
          },
          "compared_record_count": {
            "title": "Compared Record Count",
            "type": "integer"
          },
          "fixture": {
            "title": "Fixture",
            "type": "boolean"
          },
          "job_id": {
            "title": "Job Id",
            "type": "string"
          },
          "job_status": {
            "title": "Job Status",
            "type": "string"
          },
          "normalized_record_count": {
            "title": "Normalized Record Count",
            "type": "integer"
          },
          "omitted_record_count": {
            "title": "Omitted Record Count",
            "type": "integer"
          },
          "package_evidence_level": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Package Evidence Level"
          },
          "package_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Package Status"
          },
          "parent_revision_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Revision Id"
          },
          "physical_evidence_present": {
            "title": "Physical Evidence Present",
            "type": "boolean"
          },
          "revision_id": {
            "title": "Revision Id",
            "type": "string"
          },
          "revision_number": {
            "title": "Revision Number",
            "type": "integer"
          },
          "source_kind": {
            "enum": [
              "interface_fixture",
              "runtime_revision"
            ],
            "title": "Source Kind",
            "type": "string"
          },
          "source_label": {
            "title": "Source Label",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "job_id",
          "revision_id",
          "revision_number",
          "parent_revision_id",
          "title",
          "job_status",
          "source_kind",
          "source_label",
          "fixture",
          "physical_evidence_present",
          "achieved_evidence_level",
          "package_status",
          "package_evidence_level",
          "allowed_claim",
          "normalized_record_count",
          "compared_record_count",
          "omitted_record_count"
        ],
        "title": "ComparisonRevisionView",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "HardwareView": {
        "additionalProperties": false,
        "properties": {
          "gcode_uploaded": {
            "title": "Gcode Uploaded",
            "type": "boolean"
          },
          "print_started": {
            "title": "Print Started",
            "type": "boolean"
          },
          "printer_connected": {
            "title": "Printer Connected",
            "type": "boolean"
          },
          "printer_selected": {
            "title": "Printer Selected",
            "type": "boolean"
          }
        },
        "required": [
          "printer_selected",
          "printer_connected",
          "gcode_uploaded",
          "print_started"
        ],
        "title": "HardwareView",
        "type": "object"
      },
...
      "InspectionFeatureView": {
        "additionalProperties": false,
        "properties": {
          "dimensions_mm": {
            "additionalProperties": {
              "type": "number"
            },
            "title": "Dimensions Mm",
            "type": "object"
          },
          "feature_id": {
            "title": "Feature Id",
            "type": "string"
          },
          "fixture": {
            "title": "Fixture",
            "type": "boolean"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "quantity": {
            "title": "Quantity",
            "type": "integer"
          },
          "required": {
            "title": "Required",
            "type": "boolean"
          },
          "source": {
            "const": "persisted_revision_spec",
            "title": "Source",
            "type": "string"
          },
          "tolerance_mm": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tolerance Mm"
          }
        },
        "required": [
          "feature_id",
          "kind",
          "dimensions_mm",
          "quantity",
          "tolerance_mm",
          "required",
          "notes",
          "source",
          "fixture"
        ],
        "title": "InspectionFeatureView",
        "type": "object"
      },
...
      "InspectionProfileView": {
        "additionalProperties": false,
        "properties": {
          "artifact": {
            "$ref": "#/components/schemas/InspectionArtifactView"
          },
          "claim_boundary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Claim Boundary"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "profile_id": {
            "title": "Profile Id",
            "type": "string"
          },
          "profile_kind": {
            "enum": [
              "printer",
              "material",
              "process",
              "orientation"
            ],
            "title": "Profile Kind",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "values": {
            "additionalProperties": true,
            "title": "Values",
            "type": "object"
          }
        },
        "required": [
          "profile_kind",
          "artifact",
          "profile_id",
          "name",
          "status",
          "claim_boundary",
          "values"
        ],
        "title": "InspectionProfileView",
        "type": "object"
      },
...
      "InspectionView": {
        "additionalProperties": false,
        "properties": {
          "claim_boundary": {
            "title": "Claim Boundary",
            "type": "string"
          },
          "features": {
            "items": {
              "$ref": "#/components/schemas/InspectionFeatureView"
            },
            "title": "Features",
            "type": "array"
          },
          "max_json_bytes": {
            "title": "Max Json Bytes",
            "type": "integer"
          },
          "profiles": {
            "items": {
              "$ref": "#/components/schemas/InspectionProfileView"
            },
            "title": "Profiles",
            "type": "array"
          },
          "reports": {
            "items": {
              "$ref": "#/components/schemas/InspectionReportView"
            },
            "title": "Reports",
            "type": "array"
          },
          "unavailable": {
            "items": {
              "$ref": "#/components/schemas/InspectionUnavailableView"
            },
            "title": "Unavailable",
            "type": "array"
          }
        },
        "required": [
          "features",
          "reports",
          "profiles",
          "unavailable",
          "max_json_bytes",
          "claim_boundary"
        ],
        "title": "InspectionView",
        "type": "object"
      },
      "JobView": {
        "additionalProperties": false,
        "properties": {
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "job_id": {
            "title": "Job Id",
            "type": "string"
          },
          "metadata": {
            "additionalProperties": true,
            "title": "Metadata",
            "type": "object"
          },
          "request": {
            "title": "Request",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "updated_at": {
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "job_id",
          "title",
          "request",
          "status",
          "created_at",
          "updated_at",
          "metadata"
        ],
        "title": "JobView",
        "type": "object"
      },
      "PackageView": {
        "additionalProperties": false,
        "properties": {
          "allowed_claim": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Allowed Claim"
          },
          "claim_boundary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Claim Boundary"
          },
          "evidence_level": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Evidence Level"
          },
          "gcode_summary": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GcodeSummaryView"
              },
              {
                "type": "null"
              }
            ]
          },
          "hardware": {
            "$ref": "#/components/schemas/HardwareView"
          },
          "slicer": {
            "additionalProperties": true,
            "title": "Slicer",
            "type": "object"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "unresolved_warning_count": {
            "title": "Unresolved Warning Count",
            "type": "integer"
          }
        },
        "required": [
          "status",
          "evidence_level",
          "allowed_claim",
          "claim_boundary",
          "hardware",
          "gcode_summary",
          "slicer",
          "unresolved_warning_count"
        ],
        "title": "PackageView",
        "type": "object"
      },
      "RevisionComparisonResponse": {
        "additionalProperties": false,
        "properties": {
          "base": {
            "$ref": "#/components/schemas/ComparisonRevisionView"
          },
          "candidate": {
            "$ref": "#/components/schemas/ComparisonRevisionView"
          },
          "capabilities": {
            "$ref": "#/components/schemas/CapabilitiesView"
          },
          "changes": {
            "items": {
              "$ref": "#/components/schemas/ComparisonChangeView"
            },
            "title": "Changes",
            "type": "array"
          },
          "claim_boundary": {
            "title": "Claim Boundary",
            "type": "string"
          },
          "complete": {
            "title": "Complete",
            "type": "boolean"
          },
          "incomplete_value_count": {
            "title": "Incomplete Value Count",
            "type": "integer"
          },
          "max_changes": {
            "title": "Max Changes",
            "type": "integer"
          },
          "max_records_per_revision": {
            "title": "Max Records Per Revision",
            "type": "integer"
          },
          "max_value_bytes": {
            "title": "Max Value Bytes",
            "type": "integer"
          },
          "omitted_change_count": {
            "title": "Omitted Change Count",
            "type": "integer"
          },
          "relationship": {
            "enum": [
              "same_revision",
              "parent_to_child",
              "child_to_parent",
              "same_job",
              "unrelated"
            ],
            "title": "Relationship",
            "type": "string"
          },
          "returned_change_count": {
            "title": "Returned Change Count",
            "type": "integer"
          },
          "schema_version": {
            "const": "1.2.0",
            "title": "Schema Version",
            "type": "string"
          },
          "summaries": {
            "items": {
              "$ref": "#/components/schemas/ComparisonAreaSummaryView"
            },
            "title": "Summaries",
            "type": "array"
          },
          "total_change_count": {
            "title": "Total Change Count",
            "type": "integer"
          }
        },
        "required": [
          "schema_version",
          "capabilities",
          "base",
          "candidate",
          "relationship",
          "complete",
          "total_change_count",
          "returned_change_count",
          "omitted_change_count",
          "incomplete_value_count",
          "summaries",
          "changes",
          "max_changes",
          "max_value_bytes",
          "max_records_per_revision",
          "claim_boundary"
        ],
        "title": "RevisionComparisonResponse",
        "type": "object"
      },
      "RevisionDetailResponse": {
        "additionalProperties": false,
        "properties": {
          "capabilities": {
            "$ref": "#/components/schemas/CapabilitiesView"
          },
          "inspection": {
            "$ref": "#/components/schemas/InspectionView"
          },
          "job": {
            "$ref": "#/components/schemas/JobView"
          },
          "manifest_available": {
            "title": "Manifest Available",
            "type": "boolean"
          },
          "package": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PackageView"
              },
              {
                "type": "null"
              }
            ]
          },
          "revision": {
            "$ref": "#/components/schemas/RevisionView"
          },
          "schema_version": {
            "const": "1.2.0",
            "title": "Schema Version",
            "type": "string"
          },
          "source": {
            "$ref": "#/components/schemas/SourceView"
          },
          "stages": {
            "items": {
              "$ref": "#/components/schemas/StageView"
            },
            "title": "Stages",
            "type": "array"
          }
        },
        "required": [
          "schema_version",
          "capabilities",
          "source",
          "manifest_available",
          "job",
          "revision",
          "stages",
          "package",
          "inspection"
        ],
        "title": "RevisionDetailResponse",
        "type": "object"
      },
...
      "RevisionView": {
        "additionalProperties": false,
        "properties": {
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "number": {
            "title": "Number",
            "type": "integer"
          },
          "parent_revision_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Revision Id"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          },
          "revision_id": {
            "title": "Revision Id",
            "type": "string"
          },
          "spec": {
            "additionalProperties": true,
            "title": "Spec",
            "type": "object"
          }
        },
        "required": [
          "revision_id",
          "number",
          "parent_revision_id",
          "reason",
          "created_at",
          "spec"
        ],
        "title": "RevisionView",
        "type": "object"
      },
      "SourceView": {
        "additionalProperties": false,
        "properties": {
          "evidence_mode": {
            "enum": [
              "real",
              "simulated",
              "fixture",
              "unavailable",
              "mixed"
            ],
            "title": "Evidence Mode",
            "type": "string"
          },
          "fixture": {
            "title": "Fixture",
            "type": "boolean"
          },
          "kind": {
            "enum": [
              "interface_fixture",
              "runtime_revision"
            ],
            "title": "Kind",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "physical_evidence_present": {
            "title": "Physical Evidence Present",
            "type": "boolean"
          }
        },
        "required": [
          "kind",
          "label",
          "evidence_mode",
          "fixture",
          "physical_evidence_present"
        ],
        "title": "SourceView",
        "type": "object"
      },
      "StageView": {
        "additionalProperties": false,
        "properties": {
          "approvals": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Approvals",
            "type": "array"
          },
          "artifacts": {
            "items": {
              "$ref": "#/components/schemas/ArtifactView"
            },
            "title": "Artifacts",
            "type": "array"
          },
          "attempt": {
            "title": "Attempt",
            "type": "integer"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "decisions": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Decisions",
            "type": "array"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "events": {
            "items": {
              "$ref": "#/components/schemas/EventView"
            },
            "title": "Events",
            "type": "array"
          },
          "evidence_level": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Evidence Level"
          },
          "evidence_mode": {
            "enum": [
              "real",
              "simulated",
              "fixture",
              "unavailable"
            ],
            "title": "Evidence Mode",
            "type": "string"
          },
          "findings": {
            "items": {
              "$ref": "#/components/schemas/FindingView"
            },
            "title": "Findings",
            "type": "array"
          },
          "stage": {
            "title": "Stage",
            "type": "string"
          },
          "stage_run_id": {
            "title": "Stage Run Id",
            "type": "string"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "summary": {
            "title": "Summary",
            "type": "string"
          },
          "tool": {
            "$ref": "#/components/schemas/ToolView"
          }
        },
        "required": [
          "stage_run_id",
          "stage",
          "status",
          "attempt",
          "evidence_mode",
          "evidence_level",
          "tool",
          "started_at",
          "completed_at",
          "summary",
          "error_message",
          "events",
          "findings",
          "artifacts",
          "decisions",
          "approvals"
        ],
        "title": "StageView",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "This API replays persisted records. It cannot upload G-code, select a printer, heat hardware, move hardware, or start manufacturing.",
    "summary": "Read-only access to persisted fabrication evidence.",
    "title": "Ariad Fabrication Journey API",
    "version": "1.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/revision-comparison": {
      "get": {
        "operationId": "revision_comparison_api_v1_revision_comparison_get",
        "parameters": [
          {
            "in": "query",
            "name": "base_job_id",
            "required": true,
            "schema": {
              "title": "Base Job Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "base_revision_id",
            "required": true,
            "schema": {
              "title": "Base Revision Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "candidate_job_id",
            "required": true,
            "schema": {
              "title": "Candidate Job Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "candidate_revision_id",
            "required": true,
            "schema": {
              "title": "Candidate Revision Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RevisionComparisonResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Revision Comparison"
      }
    },
...
    "/api/v1/revisions/{job_id}/{revision_id}": {
      "get": {
        "operationId": "revision_api_v1_revisions__job_id___revision_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "revision_id",
            "required": true,
            "schema": {
              "title": "Revision Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RevisionDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Revision"
      }
    },
    "/api/v1/revisions/{job_id}/{revision_id}/artifacts/{artifact_id}": {
      "get": {
        "operationId": "artifact_api_v1_revisions__job_id___revision_id__artifacts__artifact_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "revision_id",
            "required": true,
            "schema": {
              "title": "Revision Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "artifact_id",
            "required": true,
            "schema": {
              "title": "Artifact Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Artifact"
      }
    }
  }
}